rename: g_data_0054206c / g_x_0054206c -> g_walkCallback
The address 0x0054206c was already declared as g_walkCallback in
include/engine/scenegraph.h (per-walk callback / scratch-during-walk
slot). 665 src/ files still referenced one of two placeholder aliases
(g_data_0054206c or g_x_0054206c) that both resolved to the same VA
via extras_map.yaml. Collapse all references to the canonical name,
drop the duplicate decls in data.c and the redundant extras_map.yaml
entries.

Header type changed from `void (*)(void)` to `unsigned int` to match
the actual usage (the slot is overloaded as a scratch integer during
the walk, so most call sites already need a cast). The four sites
that previously cast the integer to a function pointer are simplified
to drop the cast.

make matching confirms byte identity.

+++ src/boot/gated_double_setup_jmp.c
@@ -137,7 +137,7 @@ void GatedDoubleSetupJmp_00412450(void) {
     g_state_004d57ac++;
     *(unsigned int *)(g_state_004d57ac * 4) = g_cj_0054205c;
     _set_eqwt(g_scaledInit_00542044);
-    _set_x74((unsigned int)g_walkCallback);
+    _set_x74(g_walkCallback);
     _set_wc(((unsigned int)&g_data_004d5888) >> 2);
     BootChainBuildAndStep_004124c0();
 }
