+++ src/boot/small_init_helpers.c
@@ -26,5 +26,5 @@ void MStackPackedInit_004049c0(void) {
  *   ret
  */
 void ScaledStoreIdx24_00406ce0(void) {
-    *(unsigned int *)(g_scaledInit_00542044 * 4 + 0x24) = (unsigned int)g_walkCallback;
+    *(unsigned int *)(g_scaledInit_00542044 * 4 + 0x24) = g_walkCallback;
 }
